- changed to 10 byte payloads - tried some optimizations

+++ OMNI/CM_SPI/command_fifo.c
@@ -8,8 +8,8 @@
 #include <stdint.h>
 
 uint8_t command_fifo[COMMAND_FIFO_SIZE][PACKET_SIZE];	// command fifo
-uint8_t c_read_ptr = 0;									// head of fifo
-uint8_t c_write_ptr = 0;								// tail of fifo
+uint16_t c_read_ptr = 0;									// head of fifo
+uint16_t c_write_ptr = 0;								// tail of fifo
 uint8_t c_size = 0;										// number of commands in fifo
 
 // gets pointer of element to write command into
